+++ src/command_def.ml
@@ -54,7 +54,7 @@ let args =
     (* Flag to list each commands with its number *)
     +> flag "-l" no_arg
     ~aliases:["-list" ; "--list"]
-    ~doc:" Print a list of all command with its number. Useful to launch with number"
+    ~doc:" Print a list of all command with its number. Useful to launch with number. Displays a star next to next command to launch."
     (* Flag to add a command to rc file, from stdin or directly *)
     +> flag "-a" no_arg
     ~aliases:["-add" ; "--add"]
@@ -91,7 +91,7 @@ let commands =
        (* To delete command from rc file *)
        else if delete then Remove_command.run ~rc:rc_content num_cmd
        (* To print current state *)
-       else if number then State.current ()
+       else if number then State.print_current ()
        (* Reset to a value *)
        else if reset_tmp then Tmp_file.reset (Option.value ~default:0 num_cmd)
        (* Else: Run the nth command *)